% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_asTableList.R
\name{asIntensityTable}
\alias{asIntensityTable}
\title{summarize MSnExp into intensity list, intensities are filtered}
\usage{
asIntensityTable(x, ms1.noise = 100, ms1.maxPeaks = Inf,
  ms1.maxIdenticalInt = 20, ms2.noise = 30, ms2.maxPeaks = 100,
  ms2.maxIdenticalInt = 6, ...)
}
\arguments{
\item{x}{an MSnExp object}

\item{ms1.noise}{noise level of ms1 spectra, passed to \code{\link{filterSpectrum}}}

\item{ms1.maxPeaks}{maximum number of retained peaks of ms1 spectra, passed to \code{\link{filterSpectrum}}}

\item{ms1.maxIdenticalInt}{maximum number of identical intensities allowed of ms1 spectra, passed to \code{\link{filterSpectrum}}}

\item{ms2.noise}{noise level of ms2 spectra, passed to \code{\link{filterSpectrum}}}

\item{ms2.maxPeaks}{maximum number of retained peaks of ms2 spectra, passed to \code{\link{filterSpectrum}}}

\item{ms2.maxIdenticalInt}{maximum number of identical intensities allowed of ms2 spectra, passed to \code{\link{filterSpectrum}}}

\item{...}{passed to \code{\link[MSnbase]{spectrapply}} or \code{\link[BiocParallel]{bplapply}}}
}
\description{
summarize MSnExp into intensity list, intensities are filtered
}
