% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_asTableList.R
\name{asMetaTable}
\alias{asMetaTable}
\title{quick summary of MSnExp}
\usage{
asMetaTable(x)
}
\arguments{
\item{x}{an MSnExp object or an oject inheriting it.}
}
\description{
quick summary of MSnExp
}
