% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_chromPeaksPostFilter.R
\name{chromPeaksPostFilter}
\alias{chromPeaksPostFilter}
\title{post filter of chromPeaks}
\usage{
chromPeaksPostFilter(x, postfilter = c(5, 500), ...)
}
\arguments{
\item{x}{An XCMSnExp object afater peak calling}

\item{postfilter}{= c(5, 500),}

\item{...}{arguments passed bplapply}
}
\description{
post filter of chromPeaks
}
