% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_internal.R
\name{.evalPeak}
\alias{.evalPeak}
\title{Check if an XIC looks like a Guassian shape}
\usage{
.evalPeak(x)
}
\arguments{
\item{x}{a data.frame with two columns named as 'mz' and 'int' (for intensity)}
}
\description{
Check if an XIC looks like a Guassian shape
}
