% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_individualPipeline.R
\name{.indiv_chromPeaksMS2}
\alias{.indiv_chromPeaksMS2}
\title{mapping ms2 chromPeaksMS2}
\usage{
.indiv_chromPeaksMS2(x, mtab_files, itab_files,
  fun_parallel = parallel::mclapply, ...)
}
\arguments{
\item{x}{the MSnExp oject}

\item{mtab_files}{paths of meta table, often returned by function \code{indiv_peakPicking }}

\item{itab_files}{paths of the intensity table, often returned by function \code{indiv_peakPicking }}

\item{fun_parallel}{the parallel function, could be \code{mclapply} or \code{bplapply}}

\item{...}{the parameters passed to parallel function}
}
\description{
mapping ms2 chromPeaksMS2
}
