% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_individualPipeline.R
\name{.indiv_ms2map}
\alias{.indiv_ms2map}
\title{Internal function for ms2mapping}
\usage{
.indiv_ms2map(itab_files, featureMeta, valideRefs, mode, refs, peaks,
  fun_parallel, ...)
}
\description{
Internal function for ms2mapping
}
