% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_extendedChromPeak.R
\name{extendedChromPeaks}
\alias{extendedChromPeaks}
\title{extending chrom peaks table with the MS2, chromatogram (with QC)}
\usage{
extendedChromPeaks(x, mtab, itab, QC = TRUE, ...)
}
\arguments{
\item{x}{an MSnExp object}

\item{mtab}{meta table, often returned by \link{asMetaTable}}

\item{itab}{intensity table, often returned by \link{asIntensityTable}}

\item{QC}{whether the QC of EIC should be calculated}

\item{...}{other parameters passed to bplapply}
}
\description{
extending chrom peaks table with the MS2, chromatogram (with QC)
}
