% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_individualPipeline.R
\name{indiv_defineFeatures}
\alias{indiv_defineFeatures}
\title{define features for individual pipeline}
\usage{
indiv_defineFeatures(files, mtab_files, rtParam, pgParam)
}
\arguments{
\item{files}{the file paths of XCMSnExp}

\item{mtab_files}{the file path of meta tables, often returned by function \code{indiv_peakPicking}}

\item{rtParam}{the retention time adjustment parameters, passed to \code{adjustRtime}}

\item{pgParam}{the peak grouping parameter, passed to \code{groupChromPeaks}}
}
\value{
It returns an object of class \code{XCMSnExp}
}
\description{
define features for individual pipeline
}
