% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_individualPipeline.R
\name{indiv_summarizeExp}
\alias{indiv_summarizeExp}
\title{Individual pipeline for summarizeExp}
\usage{
indiv_summarizeExp(x, mode = c("pos", "neg")[1],
  peakPickingParam = MatchedFilterParam(fwhm = 7.5),
  postfilterParam = list(postfilter = c(5, 1000), ms1.noise = 100,
  ms1.maxPeaks = Inf, ms1.maxIdenticalInt = 20, ms2.noise = 30,
  ms2.maxPeaks = 100, ms2.maxIdenticalInt = 6, BPPARAM = bpparam()),
  PeakGroupsParam = PeakGroupsParam(minFraction = 0.5, extraPeaks = 1,
  smooth = "loess", span = 0.5, family = "gaussian", peakGroupsMatrix =
  matrix(nrow = 0, ncol = 0), subset = integer(), subsetAdjust =
  c("average", "previous")), retentionTimeParam = NearestPeaksParam(),
  massTab = NULL, refSpectra = NULL, parallelFun = bplapply,
  tmpdir = NA, ...)
}
\arguments{
\item{x}{a data.frame where the first columns should be named as "file". It 
contains the full path of mzXML or mzML file}

\item{mode}{mode. pos or neg}

\item{peakPickingParam}{peak picking parameters, passed to \code{findChromPeaks}}

\item{postfilterParam}{passed to \code{chromPeaksPostFilter}}

\item{PeakGroupsParam}{the peak grouping parameter, passed to \code{groupChromPeaks}}

\item{retentionTimeParam}{the retention time adjustment parameters, passed to \code{adjustRtime}}

\item{massTab}{a data.frame of mass table having at least one column named as "monoisotopic_molecular_weight"}

\item{refSpectra}{reference MS2 spectra used annotate the experimental MS2 spectra}

\item{tmpdir}{the temporary directory to store the processed results}

\item{...}{the parameters passed to parallel function}

\item{fun_parallel}{the parallel function, could be \code{mclapply} or \code{bplapply}}
}
\description{
the individual pipeline process each file separately so less memory
  used, but slower.
}
