% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_individualPipeline.R
\name{indiv_xcmsAnnotate}
\alias{indiv_xcmsAnnotate}
\title{xcms annotation function for the individual pipeline}
\usage{
indiv_xcmsAnnotate(x, mode = c("pos", "neg")[1], massTab = NULL,
  refSpectra = NULL, fun_parallel = mclapply, ...)
}
\arguments{
\item{x}{an object of XCMSnExp}

\item{mode}{mode. pos or neg}

\item{massTab}{a data.frame of mass table having at least one column named as "monoisotopic_molecular_weight"}

\item{refSpectra}{reference MS2 spectra used annotate the experimental MS2 spectra}

\item{fun_parallel}{the parallel function, could be \code{mclapply} or \code{bplapply}}

\item{...}{the parameters passed to parallel function}
}
\description{
xcms annotation function for the individual pipeline
}
