% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_individualPipeline.R
\name{indiv_xcmsSummarize}
\alias{indiv_xcmsSummarize}
\title{xcms summarize function for the individual pipeline}
\usage{
indiv_xcmsSummarize(x, mtab_files, itab_files, fun_parallel = mclapply,
  ...)
}
\arguments{
\item{x}{the MSnExp oject}

\item{mtab_files}{paths of meta table, often returned by function \code{indiv_peakPicking }}

\item{itab_files}{paths of the intensity table, often returned by function \code{indiv_peakPicking }}

\item{fun_parallel}{the parallel function, could be \code{mclapply} or \code{bplapply}}

\item{...}{the parameters passed to parallel function}
}
\description{
xcms summarize function for the individual pipeline
}
