% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_queryMass.R
\name{massQuery}
\alias{massQuery}
\title{Query possible metabolites according to given m/z}
\usage{
massQuery(m, tolppm, refTab, addTable, ID = NULL)
}
\arguments{
\item{m}{the mass of the metabolites}

\item{tolppm}{the mass tolerence given by parts per million (PPM)}

\item{refTab}{the reference metabolites table, it must contain at list the following columns:
name - the name of metabolites
monoisotopic_molecular_weight - the monoisotopic molecular weight of metabolites}

\item{addTable}{the adduct table, it must contain at least the following columns:
adduct - the name of the adduct, such as "[M+H]+"
massdiff - the mass different of an adduct
nmol - the mol of metabolite, for example, [2M-H] will be 2
ips - Four values are possible (0.25,0.5,0.75,1) depending on the likelihood of the rule}

\item{ID}{an optional ID column added at the end of the matched data.frame}
}
\description{
Query possible metabolites according to given m/z
}
