% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_spectraComparison.R
\name{spectraCosDist}
\alias{spectraCosDist}
\title{mapping MS2 spectra}
\usage{
spectraCosDist(ftid, possibleRefs, refSpectra, peakidx, peakTab, itab,
  mode = c("pos", "neg")[1])
}
\arguments{
\item{ftid}{feature id to be explored}

\item{possibleRefs}{possible references, return by annotation by mass (querymass)}

\item{refSpectra}{reference spectra, a list has two elements named as 'meta' and 'peakList',e.g. something returned by "parseHMDBMSMSXML"}

\item{peakidx}{the peak idx of that ftid}

\item{peakTab}{the peak table}

\item{itab}{the itab}

\item{mode}{either "neg" or "pos" for negative and positive mode}
}
\description{
mapping MS2 spectra
}
