% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_summarizeExp.R
\name{summarizeExp}
\alias{summarizeExp}
\title{Summarize XCMSnExp data}
\usage{
summarizeExp(x, mode = c("pos", "neg")[1], massTab = NULL,
  refSpectra = NULL, ...)
}
\arguments{
\item{x}{an object of XCMSnExp}

\item{mode}{mode. pos or neg}

\item{massTab}{a data.frame of mass table having at least one column named as "monoisotopic_molecular_weight"}

\item{refSpectra}{reference MS2 spectra used annotate the experimental MS2 spectra}

\item{...}{other parameters passed to bplapply}
}
\description{
Summarize XCMSnExp data
}
